+++ db/migrate/20210625143438_add_mobile_number_to_claims.rb
@@ -1,0 +1,5 @@
+class AddMobileNumberToClaims < ActiveRecord::Migration[6.0]
+  def change
+    add_column :claims, :mobile_number, :string
+  end
+end