+++ spec/helpers/claims_helper_spec.rb
@@ -90,6 +90,7 @@ describe ClaimsHelper do
 
       it "returns an array of identity-related questions and answers for displaying to the user for review" do
         claim.provide_mobile_number = "Yes"
+        claim.mobile_number = "01234567899"
 
         expected_answers = [
           [I18n.t("questions.name"), "Jo Bloggs", "personal-details"],
@@ -99,7 +100,8 @@ describe ClaimsHelper do
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
           [I18n.t("questions.national_insurance_number"), "QQ123456C", "personal-details"],
           [I18n.t("questions.email_address"), "test@email.com", "email-address"],
-          [I18n.t("questions.provide_mobile_number"), "Yes", "provide-mobile-number"]
+          [I18n.t("questions.provide_mobile_number"), "Yes", "provide-mobile-number"],
+          [I18n.t("questions.mobile_number"), "01234567899", "mobile-number"]
         ]
 
         expect(helper.identity_answers(claim)).to eq expected_answers
@@ -116,7 +118,8 @@ describe ClaimsHelper do
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
           [I18n.t("questions.national_insurance_number"), "QQ123456C", "personal-details"],
           [I18n.t("questions.email_address"), "test@email.com", "email-address"],
-          [I18n.t("questions.provide_mobile_number"), "No", "provide-mobile-number"]
+          [I18n.t("questions.provide_mobile_number"), "No", "provide-mobile-number"],
+          [I18n.t("questions.mobile_number"), nil, "mobile-number"]
         ]
 
         expect(helper.identity_answers(claim)).to eq expected_answers